+++ src/main/ipc/projectHandlers.ts
@@ -8,5 +8,6 @@ export {
   handleLaunchProjectWithConfig
 } from './projectLaunching'
 export { calculateProjectSize, calculateAllProjectSizes } from '../utils/projectSizing'
-export { scanAndMergeProjects, loadSavedProjects, deleteProject, updateProjectVersion } from '../utils/projectValidation'
+export { scanAndMergeProjects, loadSavedProjects, deleteProject, eraseProjectFromDisk, updateProjectVersion } from '../utils/projectValidation'
 export { checkProjectHealth } from '../utils/projectHealth'
+